# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trust Guard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardStatsRiskDistribution = {
  tier1: number;
  tier2: number;
  tier3: number;
  tier4: number;
  tier5: number;
  tier6: number;
};

export interface LogEntry {
  id: number;
  userId: string;
  guildId: string;
  username: string;
  eventType: string;
  metadata?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalVerified: number;
  totalFailed: number;
  pendingReviews: number;
  dailyJoins: number;
  riskDistribution: DashboardStatsRiskDistribution;
  recentActivity: LogEntry[];
}

export interface VerifiedUserSummary {
  id: string;
  userId: string;
  guildId: string;
  username: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  riskScore: number;
  verificationTier: number;
  status: string;
  verifiedAt?: string | null;
  joinedAt: string;
  accountCreatedAt: string;
  hasAvatar: boolean;
  hasBanner: boolean;
  attemptCount: number;
  createdAt: string;
}

export interface UserListResponse {
  users: VerifiedUserSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AttemptSummary {
  id: number;
  tier: number;
  riskScore: number;
  status: string;
  captchaPassed?: boolean | null;
  questionnairePassed?: boolean | null;
  challengePassed?: boolean | null;
  failureReason?: string | null;
  startedAt: string;
  completedAt?: string | null;
}

export interface UserDetail {
  user: VerifiedUserSummary;
  attempts: AttemptSummary[];
  logs: LogEntry[];
}

export interface ReviewSummary {
  id: number;
  userId: string;
  guildId: string;
  username: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  riskScore: number;
  tier: number;
  questionnaireResponses?: string | null;
  status: string;
  staffNotes?: string | null;
  createdAt: string;
  resolvedAt?: string | null;
}

export interface ReviewListResponse {
  reviews: ReviewSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ReviewDecisionBody {
  notes?: string;
}

export interface ReviewDecisionResponse {
  success: boolean;
  message: string;
}

export interface LogListResponse {
  logs: LogEntry[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetDashboardUsersParams = {
page?: number;
limit?: number;
search?: string;
status?: string;
tier?: number;
};

export type GetPendingReviewsParams = {
page?: number;
limit?: number;
};

export type GetVerificationLogsParams = {
page?: number;
limit?: number;
userId?: string;
eventType?: string;
};