# lib/api-zod/src/generated/types/reviewDecisionResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trust Guard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewDecisionResponse {
  success: boolean;
  message: string;
}