# lib/api-zod/src/generated/types/reviewListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trust Guard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewSummary } from './reviewSummary';

export interface ReviewListResponse {
  reviews: ReviewSummary[];
  total: number;
  page: number;
  totalPages: number;
}