# lib/api-zod/src/generated/types/reviewSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trust Guard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewSummary {
  id: number;
  userId: string;
  guildId: string;
  username: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  riskScore: number;
  tier: number;
  questionnaireResponses?: string | null;
  status: string;
  staffNotes?: string | null;
  createdAt: string;
  resolvedAt?: string | null;
}